added src/main.cpp, bash/burgers1d.sh

+++ include/burgers.hpp
@@ -1,6 +1,6 @@
 int inviscidBurgers1d(const double* u_0x,
-                      const double& nt,
-                      const double& nx,
+                      const int& nt,
+                      const int& nx,
                       const double& dt,
                       const double& dx,
                       const double& tol,
